+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapter.java
@@ -145,16 +145,27 @@ public class ExoPlayerAdapter implements PlayerAdapter, EventDataManipulator {
     */
     private void checkAutoplayStartup() {
         int playbackState = exoplayer.getPlaybackState();
-        boolean playWhenReady = exoplayer.getPlayWhenReady();
-        if (playbackState == Player.STATE_BUFFERING) {
-            if (playWhenReady != this.isPlaying && playWhenReady) {
+
+        boolean isBufferingAndWillAutoPlay =
+                exoplayer.getPlayWhenReady() && playbackState == Player.STATE_BUFFERING;
+        /* Even if flag was set as `player.setPlayWhenReady(false)`, when player is playing, flags is returned as `true` */
+        boolean isAlreadyPlaying =
+                exoplayer.getPlayWhenReady() && playbackState == Player.STATE_READY;
+
+        if (isBufferingAndWillAutoPlay || isAlreadyPlaying) {
+            this.isPlaying = true;
+
+            long position = getPosition();
+            Log.d(
+                    TAG,
+                    "Collector was attached while media source was already loading, transitioning to startup state.");
+            startup(position);
+
+            if (playbackState == Player.STATE_READY) {
                 Log.d(
                         TAG,
-                        "Collector was attached while media source was already loading, transitioning to startup state.");
-                // with autoplay enabled the player first enter here and start buffering for the
-                // video with playWhenReady = true
-                this.isPlaying = true;
-                startup(getPosition());
+                        "Collector was attached while media source was already playing, transitioning to playing state");
+                stateMachine.transitionState(PlayerState.PLAYING, position);
             }
         }
     }